+++ pages/signup.tsx
@@ -1,4 +1,4 @@
-import Signup from "../components/Signup"
+import Signup from "../components/Signup/Signup"
 
 const Home = () => {
     return (
